Move `TsrgFileReader` into `srg` package
